refactored grow loop and use redis to serve embeddings
