+++ codescholar/search/search.py
@@ -1,4 +1,5 @@
 import os
+import redis
 import os.path as osp
 import argparse
 import numpy as np
@@ -20,27 +21,25 @@ from codescholar.sast.visualizer import render_sast
 from codescholar.sast.sast_utils import sast_to_prog
 from codescholar.representation import models, config
 from codescholar.search import search_config
+from codescholar.search.grow import grow
 from codescholar.search.elastic_search import grep_programs
 from codescholar.search.init_search import init_search_q, init_search_m, init_search_mq
 from codescholar.utils.search_utils import (
-    sample_programs,
     ping_elasticsearch,
     ping_elasticindex,
     wl_hash,
     save_idiom,
     _print_mine_logs,
     _write_mine_logs,
-    _frontier,
-    read_graph,
-    read_prog,
-    read_embedding,
-    read_embeddings_batched,
+    read_graph
 )
+from codescholar.utils.search_utils import read_embedding, save_embedding_to_redis
 from codescholar.utils.train_utils import build_model, get_device, featurize_graph
 from codescholar.utils.graph_utils import nx_to_program_graph, program_graph_to_nx
 from codescholar.utils.perf import perftimer
 from codescholar.constants import DATA_DIR
 
+redis_client = redis.StrictRedis(host='localhost', port=6379, db=0)
 
 ######### IDIOM STORE ############
 
@@ -119,151 +118,10 @@ def start_workers_grow(prog_indices, in_queue, out_queue, args, gpu):
 
     return workers
 
-
-def score_candidate_freq(args, model, embs, cand_emb, device_id=None):
-    """score candidate program embedding against target program embeddings
-    in a batched manner.
-
-    Algorithm:
-    softmax based classifier on top of emb-diff
-        score = #nhoods !containing cand.
-        = count(cand_emb - embs > 0)
-            --> classifier: 0/1
-    """
-    score = 0
-
-    if cand_emb.shape[0] > 1:
-        preds = []
-
-        for comp_emb in cand_emb:
-            comp_preds = np.array([])
-            for emb_batch in embs:
-                with torch.no_grad():
-                    is_subgraph_rel = model.predict((emb_batch.to(get_device(device_id)), comp_emb))
-                    is_subgraph = model.classifier(is_subgraph_rel.unsqueeze(1))
-                    comp_preds = np.concatenate((comp_preds, torch.argmax(is_subgraph, axis=1).cpu().numpy()))
-
-            preds.append(comp_preds)
-
-        assert len(set([len(pred) for pred in preds])) == 1, "component preds have different shapes!"
-
-        preds = np.array(preds)
-        merged_preds = np.all(preds, axis=0)
-        score = np.sum(merged_preds)
-    else:
-        for emb_batch in embs:
-            with torch.no_grad():
-                is_subgraph_rel = model.predict((emb_batch.to(get_device(device_id)), cand_emb))
-                is_subgraph = model.classifier(is_subgraph_rel.unsqueeze(1))
-                score += torch.sum(torch.argmax(is_subgraph, axis=1)).item()
-
-    return score
-
-
-def grow(args, prog_indices, in_queue, out_queue, device_id=None):
-    embs = read_embeddings_batched(args, prog_indices)
-
-    codebert_name = "microsoft/codebert-base"
-    feat_tokenizer = RobertaTokenizer.from_pretrained(codebert_name)
-    feat_model = RobertaModel.from_pretrained(codebert_name).to(get_device(device_id))
-    feat_model.eval()
-
-    # print("Moving model to device:", get_device(device_id))
-    model = build_model(models.SubgraphEmbedder, args, device_id=device_id)
-    model.eval()
-
-    done = False
-    while not done:
-        msg, beam_set = in_queue.get()
-
-        if msg == "done":
-            del embs
-            done = True
-            break
-
-        new_beams = []
-
-        # STEP 1: Explore all candidates in each beam of the beam_set
-        for beam in beam_set:
-            _, holes, neigh, frontier, visited, graph_idx = beam
-            graph = read_graph(args, graph_idx)
-
-            if len(neigh) >= args.max_idiom_size or not frontier:
-                continue
-
-            cand_neighs = []
-
-            # EMBED CANDIDATES
-            for i, cand_node in enumerate(frontier):
-                cand_neigh = graph.subgraph(neigh + [cand_node])
-                connected_comps = list(nx.connected_components(cand_neigh.to_undirected()))
-
-                if len(connected_comps) == 1:
-                    cand_neigh = featurize_graph(cand_neigh, feat_tokenizer, feat_model, anchor=neigh[0], device_id=device_id)
-                    cand_neighs.append(cand_neigh)
-                else:
-                    comp_neighs = []
-                    for comp in connected_comps:
-                        comp_neigh = cand_neigh.subgraph(comp)
-                        comp_root = [n for n in comp_neigh.nodes if comp_neigh.in_degree(n) == 0][0]
-                        comp_neigh = featurize_graph(comp_neigh, feat_tokenizer, feat_model, anchor=comp_root, device_id=device_id)
-                        comp_neighs.append(comp_neigh)
-
-                    cand_neighs.append(comp_neighs)
-
-            flat_cand_neighs = list(chain.from_iterable([x if isinstance(x, list) else [x] for x in cand_neighs]))
-            cand_batch = Batch.from_data_list(flat_cand_neighs).to(get_device(device_id))
-
-            with torch.no_grad():
-                cand_embs = model.encoder(cand_batch)
-
-            # SCORE CANDIDATES (freq)
-            for i, (cand_node, cand_neigh) in enumerate(zip(frontier, cand_neighs)):
-                if isinstance(cand_neigh, list):
-                    cand_emb = cand_embs[i : i + len(cand_neigh)]
-                else:
-                    cand_emb = cand_embs[i : i + 1]
-
-                # first, add new holes introduced
-                # then, remove hole filled in/by cand_node (incoming/outgoing edge resp)
-                new_holes = holes + graph.nodes[cand_node]["span"].count("#") - 1
-
-                # filter out candidates that exceed max_holes
-                if new_holes < 0 or new_holes > args.max_holes:
-                    continue
-
-                new_neigh = neigh + [cand_node]
-
-                # new frontier = {prev frontier} U {outgoing and incoming neighbors of cand_node} - {visited}
-                # note: one can use type='neigh' to add only outgoing neighbors
-                new_frontier = list(((set(frontier) | _frontier(graph, cand_node, type="radial")) - visited) - set([cand_node]))
-
-                new_visited = visited | set([cand_node])
-
-                score = score_candidate_freq(args, model, embs, cand_emb, device_id=device_id)
-                new_beams.append((score, new_holes, new_neigh, new_frontier, new_visited, graph_idx))
-
-        # STEP 2: Sort new beams by freq_score/#holes
-        new_beams = list(sorted(new_beams, key=lambda x: x[0] / x[1] if x[1] > 0 else x[0], reverse=True))
-
-        # print("===== [debugger] new beams =====")
-        # for beam in new_beams:
-        #     print("freq: ", beam[0])
-        #     print("holes: ", beam[1])
-        #     print("nodes: ", [graph.nodes[n]['span'] for n in beam[2]])
-        #     print("frontier: ", [graph.nodes[n]['span'] for n in beam[3]])
-        #     print()
-        # print("================================")
-
-        # STEP 3: filter top-k beams
-        new_beams = new_beams[: args.n_beams]
-        out_queue.put(("complete", new_beams))
-
-
 ######### MAIN ############
 
 
-@perftimer
+# @perftimer
 def search(args, prog_indices, beam_sets):
     mine_summary = defaultdict(lambda: defaultdict(int))
     size = 1
@@ -376,6 +234,10 @@ def main(args):
         beam_sets = init_search_m(args, prog_indices)
     else:
         raise ValueError(f"Invalid search mode {args.mode}!")
+
+    # load all embeddings of prog_indices to redis
+    for idx in tqdm(prog_indices, desc="Load embeddings"):
+        save_embedding_to_redis(idx, read_embedding(args, idx))
 
     # STEP 2: search for idioms; saves idioms gradually
     mine_summary = search(args, prog_indices, beam_sets)
